requirements.txt
requests
beautifulsoup4
pandas
openai
tldextract
python-dotenv
gspread
oauth2client

